tested inputs
